Comments, yes/no terminal query, flag added to DirContents
Added some comments to explain what is going on in some places.
Function to query the user in terminal and return a boolean.
Added a flag to DirContents to be used with the getDirContents
function in order to indicate that the struct is a representation
of the directory that was received as a parameter to the function